app/stacks/dataplatform/processing/glue-jobs/requirements.txt
pyspark
apache-sedona==1.7.0
awsglue-dev